+++ src/main/java/yesman/epicfight/client/events/engine/RenderEngine.java
@@ -372,6 +372,23 @@ public class RenderEngine implements IEventBasedEngine {
 	public static boolean hitResultNotEquals(@Nullable HitResult hitResult, HitResult.Type hitType) {
 		return hitResult == null ? true : !hitType.equals(hitResult.getType());
 	}
+
+    /**
+     * @deprecated Left for Shoulder Surfing Backward-Compatibility.
+     * <a href="https://github.com/Exopandora/ShoulderSurfing/issues/359#issuecomment-3551814165">Related issue</a>
+     */
+    @Deprecated(forRemoval = true)
+    public void correctCamera(ViewportEvent.ComputeCameraAngles event, float partialTicks) {
+        event.getCamera().setRotation(1.0F, 1.0F);
+    }
+
+    /**
+     * @deprecated Left for Shoulder Surfing Backward-Compatibility.
+     * <a href="https://github.com/Exopandora/ShoulderSurfing/issues/359#issuecomment-3551814165">Related issue</a>
+     */
+    @Deprecated(forRemoval = true)
+    public void setRangedWeaponThirdPerson(ViewportEvent.ComputeCameraAngles event, CameraType pov, double partialTicks) {
+    }
 	
 	/******************
 	 * Forge Event listeners