+++ src/main/java/yesman/epicfight/compat/shouldersurfing/ShoulderSurfingCompat.java
@@ -10,6 +10,8 @@ import yesman.epicfight.api.client.input.handlers.InputManager;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.api.distmarker.OnlyIn;
 import yesman.epicfight.client.ClientEngine;
+import yesman.epicfight.client.camera.EpicFightTpsCameraDisableState;
+import yesman.epicfight.client.camera.EpicFightTpsCameraDisabledReason;
 import yesman.epicfight.client.world.capabilites.entitypatch.player.LocalPlayerPatch;
 
 /**
@@ -48,6 +50,7 @@ public class ShoulderSurfingCompat implements IShoulderSurfingPlugin {
     public void register(IShoulderSurfingRegistrar registrar) {
         registrar.registerCameraCouplingCallback(new ForceCameraCouplingWhenAttackingCallback());
         registrar.registerCameraCouplingCallback(new ForceCameraCouplingWhenHoldingSkillCallback());
+        EpicFightTpsCameraDisableState.disable(EpicFightTpsCameraDisabledReason.ShoulderSurfing);
     }
 
     private static class ForceCameraCouplingWhenAttackingCallback implements ICameraCouplingCallback {